more debug info (smart dashboard for if tilt correction is happening)

+++ FRC2018/src/PenguinEmpire.h
@@ -132,8 +132,8 @@ public:
 
 	int dist;
 
-	int leftXVision;
-	int rightXVision;
+	double leftXVision;
+	double rightXVision;
 
 // Stages
